requirements.txt
astroid==1.6.1
colorama==0.3.9
Django==2.0.2
django-markdown-deux==1.0.5
django-pagedown==1.0.4
isort==4.3.4
lazy-object-proxy==1.3.1
markdown2==2.3.5
mccabe==0.6.1
mysqlclient==1.3.12
Pillow==5.0.0
pylint==1.8.2
pytz==2018.3
six==1.11.0
wrapt==1.10.11
